Adjust drawing code for compatibility with CloseView.
CloseView allocates memory for the screen at a different location from the hardware video memory, then mirrors that to the actual video memory (possibly magnified). Writing directly to video memory as Prizm does will lead to screen corruption. To avoid this, we now detect when the screen has been moved like this (as discussed in IIGS Tech Note #91) and fall back to the slower QuickDraw-based drawing code in this case. The code to draw the debugger icons in the top right is also adjusted.

With these changes, Prizm should now work correctly if CloseView is installed but turned off. It also mostly works while CloseView is active, but there can still be issues if a desktop program is debugged while the screen is magnified.